feat:Update to modules file structure
