+++ submissions/week-2/day-3/Samson-Aderonmu/garage_access_control/src/state.rs
@@ -1,0 +1,15 @@
+#[derive(Debug, PartialEq, Clone)]
+pub enum EmployeeType {
+    MediaTeam,
+    IT,
+    Manager,
+    SocialMedia,
+    TechnicianSupervisor,
+    KitchenStaff,
+}
+
+#[derive(Clone)]
+pub struct Employee {
+    pub employee_type: EmployeeType,
+    pub is_employed: bool,
+}
